download frames images when uploading a video file

+++ zebrafish-data-pipeline-be/app.py
@@ -1,6 +1,7 @@
-from flask import Flask, request, jsonify
-import os, base64
+from flask import Flask, request, jsonify, send_file
+import os, base64, zipfile
 from flask_cors import CORS  # Import the CORS extension
+from io import BytesIO
 
 
 
@@ -15,7 +16,7 @@ def extract_images():
     # Get the uploaded videos from the request
     #data = request.get_json()
     #print("RECEIVED DATA: ",data)
-    data_path =  "/Users/ambroseling/Desktop/zebrafish-data-pipeline/data"
+    data_path =  "../data"
     uploaded_files = request.files.getlist('videos')
     print("UPLOADED FILES:")
     print(uploaded_files)
@@ -54,7 +55,7 @@ def extract_images():
 
 @app.route('/get_images', methods=['GET'])
 def get_images():
-    image_directory = '/Users/ambroseling/Desktop/zebrafish-data-pipeline/data'
+    image_directory = '../data'
     image_data_list = []
 
     # Iterate through the image files in the directory
@@ -71,10 +72,32 @@ def get_images():
 
             # Append the base64-encoded image data to the list
             image_data_list.append({'filename': filename, 'data': base64_image})
-            os.remove(image_path)
+            # os.remove(image_path)
 
     # Return the list of base64-encoded image data in the JSON response
     return jsonify({'images': image_data_list})
+
+
+@app.route('/download_images', methods=['GET'])
+def download_images():
+    image_directory = '../data'
+
+    memory_file = BytesIO()
+    with zipfile.ZipFile(memory_file, 'w', zipfile.ZIP_DEFLATED) as zipf:
+        for filename in os.listdir(image_directory):
+            if filename.endswith('.jpg') or filename.endswith('.png'):
+                image_path = os.path.join(image_directory, filename)
+                zipf.write(image_path, os.path.relpath(image_path, image_directory))
+                os.remove(image_path)
+
+    memory_file.seek(0)
+    return send_file(
+        memory_file,
+        mimetype='application/zip',
+        as_attachment=True,
+        download_name='extracted_frames.zip'
+    )
+
 
 if __name__ == '__main__':
     app.run(debug=True)
